src/static/requirements.txt
bs4
jsbeautifier
jsoncomment
jstyleson
jsmin
psycopg
psycopg_binary
psycopg2-binary
psycopg-pool
pycparser
retirejs
sqlparse
tldextract
tqdm
typing_extensions
